Move MakeWorkDir to executil, add Dir cmd helper
