+++ executil/executil_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-package buildmodel
+package executil
 
 import (
 	"path"
